Refactor ScaffoldedQuestion to use BaseQuestionApp
[#172821069]
